Fix compiler warning

+++ src/models/itemmodel.cpp
@@ -64,10 +64,10 @@ QHash<int, QByteArray> ItemModel::roleNames() const
  */
 void ItemModel::setProbability(const int row, const float probability)
 {
-	if (row < 0 || row >= itemConfigs.size() || probability < 0 || probability > 1) {
+	if (row < 0 || static_cast<size_t>(row) >= itemConfigs.size() || probability < 0 || probability > 1) {
 		return;
 	}
-	itemConfigs[static_cast<unsigned long>(row)].probability = probability;
+	itemConfigs[static_cast<size_t>(row)].probability = probability;
 }
 
 /**
